refactor(srv): extract main()'s 1165-line bootstrap sequence into bootstrap.rs
Tier 3 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
async fn main() was a single flat, numbered bootstrap sequence (crash-monitor
branch, process watchers, logging, config, DB/migration, background
subsystems, listener binds, AppState assembly, shutdown handlers). Extracted
each phase into a named bootstrap:: function threaded through small bundle
structs (Stores/BackgroundSubsystems/NetworkBundle/ReducerPlumbing); main()
is now a ~178-line linear orchestrator calling them in the same order.

No logic changes — every relocated block is byte-for-byte identical, only
wrapped in named functions. main.rs: 1509 -> 178 lines; bootstrap.rs: 1593
lines (new).

cargo build -p agentmux-srv clean; cargo test -p agentmux-srv --bin
agentmux-srv — 1640 passed, 0 failed, 4 ignored; cargo test -p agentmux-srv
--test integration_test — 4/4 pass, including sigterm_exits_process which
exercises the new install_shutdown_handlers path end-to-end against the
real spawned binary.

<!-- agentmux:agent_id=agentx -->
